requirements.txt
requests>=2.28.2
websocket-client>=1.5.1
python-socketio[client]>=5.8.0
numpy>=1.24.2
opencv-python>=4.6.0.66
types-requests==2.31.0.1
era-5g-interface~=0.6.0
ujson~=5.8.0
types-ujson~=5.8.0.1
flask_socketio==5.3.*
flask==2.2.*